introduce MainWindow
