+++ src/main/sessions.ts
@@ -3,11 +3,7 @@
 
 import {
   app,
-  BrowserWindow,
-  clipboard,
   ipcMain,
-  Menu,
-  MenuItemConstructorOptions
 } from 'electron';
 
 import { JSONObject } from '@lumino/coreutils';
@@ -29,8 +25,7 @@ import { appConfig } from './utils';
 import { EventEmitter } from 'events';
 
 import * as fs from 'fs';
-import { LabView } from './labview/labview';
-import { TitleBarView } from './titlebarview/titlebarview';
+import { MainWindow } from './mainwindow/mainwindow';
 
 export interface ISessions extends EventEmitter {
   createSession: (opts?: JupyterLabSession.IOptions) => Promise<void>;
@@ -422,85 +417,19 @@ export class JupyterLabSession {
       }
     }
 
-    const labView = new LabView({
-      serverState: this._info.serverState,
-      platform: this._info.platform,
-      uiState: this._info.uiState
-    });
+    this._window = new MainWindow(this._info);
 
-    this._window = new BrowserWindow({
-      width: this._info.width,
-      height: this._info.height,
-      x: this._info.x,
-      y: this._info.y,
-      minWidth: 400,
-      minHeight: 300,
-      show: true,
-      title: 'JupyterLab',
-      titleBarStyle: 'hidden',
-      webPreferences: {
-        nodeIntegration: true,
-        contextIsolation: false
-      }
-    });
-
-    this._window.setMenuBarVisibility(false);
-    this._addFallbackContextMenu();
-
-    if (this._info.x && this._info.y) {
-      this._window.setBounds({
-        x: this._info.x,
-        y: this._info.y,
-        height: this._info.height,
-        width: this._info.width
-      });
-    } else {
-      this._window.center();
-    }
-
-    this._addRenderAPI();
-
-    this._window.on('focus', () => {
+    this._window.window.on('focus', () => {
       this._sessionManager.setFocusedSession(this);
     });
 
     sessionManager.app.pageConfigSet.then(() => {
-      const titleBarView = new TitleBarView();
-      this._window.addBrowserView(titleBarView.view);
-      titleBarView.view.setBounds({ x: 0, y: 0, width: 1200, height: 100 });
-      titleBarView.load();
-
-      this._window.addBrowserView(labView.view);
-      labView.view.setBounds({ x: 0, y: 100, width: 1200, height: 700 });
-      labView.load();
-
-      const resizeViews = () => {
-        const titleBarHeight = 28;
-        const [width, height] = this._window.getSize();
-        titleBarView.view.setBounds({
-          x: 0,
-          y: 0,
-          width: width,
-          height: titleBarHeight
-        });
-        labView.view.setBounds({
-          x: 0,
-          y: titleBarHeight,
-          width: width,
-          height: height - titleBarHeight
-        });
-      };
-
-      this._window.on('resize', () => {
-        resizeViews();
-      });
-
-      resizeViews();
+      this._window.load();
     });
   }
 
   get info(): JupyterLabSession.IInfo {
-    let winBounds = this._window.getBounds();
+    let winBounds = this._window.window.getBounds();
     this._info.x = winBounds.x;
     this._info.y = winBounds.y;
     this._info.width = winBounds.width;
@@ -509,7 +438,7 @@ export class JupyterLabSession {
   }
 
   get browserWindow(): Electron.BrowserWindow {
-    return this._window;
+    return this._window.window;
   }
 
   state(): JupyterLabSession.IState {
@@ -525,118 +454,11 @@ export class JupyterLabSession {
     };
   }
 
-  /**
-   * Simple fallback context menu shown on Shift + Right Click.
-   * May be removed in future versions once (/if) JupyterLab builtin menu
-   * supports cut/copy/paste, including "Copy link URL" and "Copy image".
-   * @private
-   */
-  private _addFallbackContextMenu(): void {
-    const selectionTemplate: MenuItemConstructorOptions[] = [{ role: 'copy' }];
-
-    const inputMenu = Menu.buildFromTemplate([
-      { role: 'cut' },
-      { role: 'copy' },
-      { role: 'paste' },
-      { role: 'selectAll' }
-    ]);
-
-    this._window.webContents.on('context-menu', (event, params) => {
-      if (params.isEditable) {
-        inputMenu.popup({ window: this._window });
-      } else {
-        const template: MenuItemConstructorOptions[] = [];
-        if (params.selectionText) {
-          template.push(...selectionTemplate);
-        }
-        if (params.linkURL) {
-          template.push({
-            label: 'Copy link URL',
-            click: () => {
-              clipboard.writeText(params.linkURL);
-            }
-          });
-        }
-        if (params.hasImageContents) {
-          template.push({
-            label: 'Copy image',
-            click: () => {
-              this._window.webContents.copyImageAt(params.x, params.y);
-            }
-          });
-        }
-        if (template.length) {
-          Menu.buildFromTemplate(template).popup({ window: this._window });
-        }
-      }
-    });
-  }
-
-  private _addRenderAPI(): void {
-    ipcMain.on('state-update', (evt: any, arg: any) => {
-      for (let key in arg) {
-        if ((this._info as any)[key]) {
-          (this._info as any)[key] = (arg as any)[key];
-        }
-      }
-    });
-
-    if (this.info.uiState === 'mac') {
-      this._window.on('enter-full-screen', () => {
-        asyncRemoteMain.emitRemoteEvent(
-          ISessions.enterFullScreenEvent,
-          undefined,
-          this._window.webContents
-        );
-      });
-
-      this._window.on('leave-full-screen', () => {
-        asyncRemoteMain.emitRemoteEvent(
-          ISessions.leaveFullScreenEvent,
-          undefined,
-          this._window.webContents
-        );
-      });
-    } else {
-      this._window.on('maximize', () => {
-        asyncRemoteMain.emitRemoteEvent(
-          ISessions.enterFullScreenEvent,
-          undefined,
-          this._window.webContents
-        );
-      });
-
-      this._window.on('unmaximize', () => {
-        asyncRemoteMain.emitRemoteEvent(
-          ISessions.leaveFullScreenEvent,
-          undefined,
-          this._window.webContents
-        );
-      });
-    }
-
-    this._window.on('minimize', () => {
-      asyncRemoteMain.emitRemoteEvent(
-        ISessions.minimizeEvent,
-        undefined,
-        this._window.webContents
-      );
-    });
-
-    this._window.on('restore', () => {
-      asyncRemoteMain.emitRemoteEvent(
-        ISessions.restoreEvent,
-        undefined,
-        this._window.webContents
-      );
-    });
-  }
-
   private _sessionManager: JupyterLabSessions = null;
 
   private _info: JupyterLabSession.IInfo = null;
 
-  private _window: Electron.BrowserWindow = null;
+  private _window: MainWindow = null;
 }
 
 export namespace JupyterLabSession {
